requirements.txt
alembic==1.11.2
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-DotEnv==0.1.2
Flask-JWT==0.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==1.4.2
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
